pyproject.toml
[project]
name = "agent-trader"
version = "0.1.0"
description = "Trump post trading agent backtest system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.44",
    "pydantic>=2.10",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "fire>=0.7",
    "rich>=13.0",
    "weave>=0.51",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "loguru>=0.7",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_trader"]

[tool.mypy]
python_version = "3.12"

[[tool.mypy.overrides]]
module = [
    "pydantic_settings",
    "loguru",
    "claude_agent_sdk",
    "jinja2",
    "rich.*",
    "weave",
    "fire",
    "dotenv",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pandas-stubs>=3.0.0.260204",
]
